delete, flag and vote comments updated app icon to hopefully be good for fastlane

+++ mist-ios/Model/Post/Vote/VoteService.swift
@@ -15,7 +15,8 @@ class VoteService: NSObject {
     
     static var singleton = VoteService()
     
-    private var votes: [PostVote] = []
+    private var postVotes: [PostVote] = []
+    private var commentVotes: [CommentVote] = []
 
     //MARK: - Initialization
     
@@ -26,84 +27,133 @@ class VoteService: NSObject {
     //MARK: - Load
     
     func loadVotes() async throws {
-        votes = try await PostVoteAPI.fetchVotesByUser(voter: UserService.singleton.getId())
+        async let loadedPostVotes = PostVoteAPI.fetchVotesByUser(voter: UserService.singleton.getId())
+        async let loadedCommentVotes = CommentVoteAPI.fetchVotesByUser(voter: UserService.singleton.getId())
+        (postVotes, commentVotes) = try await (loadedPostVotes, loadedCommentVotes)
     }
     
     //MARK: - Getters
     
     func votesForPost(postId: Int) -> [PostVote] {
-        return votes.filter { $0.post == postId }
+        return postVotes.filter { $0.post == postId }
     }
     
-    //MARK: - Updaters
+    func voteForComment(commentId: Int) -> CommentVote? {
+        return commentVotes.filter { $0.comment == commentId }.first
+    }
+    
+    //MARK: - Post Vote Updaters
     
     // Intermediate layer
-    func handleVoteUpdate(postId: Int, emoji: String, _ action: VoteAction) throws {
+    func handlePostVoteUpdate(postId: Int, emoji: String, _ action: VoteAction) throws {
         switch action {
         case .cast:
-            try castVote(postId: postId, emoji: emoji)
+            try castPostVote(postId: postId, emoji: emoji)
         case .patch:
-            try patchVote(postId: postId, emoji: emoji)
+            try patchPostVote(postId: postId, emoji: emoji)
         case .delete:
-            try deleteVote(postId: postId)
+            try deletePostVote(postId: postId)
         }
     }
     
-    private func castVote(postId: Int, emoji: String) throws {
+    private func castPostVote(postId: Int, emoji: String) throws {
         let addedVote = PostVote(id: Int.random(in: 0..<Int.max),
                                  voter: UserService.singleton.getId(),
                                  post: postId,
                                  timestamp: Date().timeIntervalSince1970,
                                  emoji: emoji,
                                  rating: nil)
-        votes.append(addedVote)
+        postVotes.append(addedVote)
         
         Task {
             do {
                 let _ = try await PostVoteAPI.postVote(voter: UserService.singleton.getId(), post: postId, emoji: emoji)
             } catch {
-                votes.removeAll { $0.id == addedVote.id }
+                postVotes.removeAll { $0.id == addedVote.id }
                 throw(error)
             }
         }
     }
     
-    private func patchVote(postId: Int, emoji: String) throws {
-        let originalVote = votes.first { $0.post == postId }!
-        votes.removeAll { $0.id == originalVote.id }
+    private func patchPostVote(postId: Int, emoji: String) throws {
+        let originalVote = postVotes.first { $0.post == postId }!
+        postVotes.removeAll { $0.id == originalVote.id }
         let patchedVote = PostVote(id: Int.random(in: 0..<Int.max),
                                  voter: UserService.singleton.getId(),
                                  post: postId,
                                  timestamp: Date().timeIntervalSince1970,
                                  emoji: emoji,
                                  rating: nil)
-        votes.append(patchedVote)
+        postVotes.append(patchedVote)
         
         Task {
             do {
                 let _ = try await PostVoteAPI.patchVote(voter: UserService.singleton.getId(), post: postId, emoji: emoji)
             } catch {
                 print(error.localizedDescription)
-                votes.removeAll { $0.id == patchedVote.id }
-                votes.append(originalVote)
+                postVotes.removeAll { $0.id == patchedVote.id }
+                postVotes.append(originalVote)
                 throw(error)
             }
         }
     }
     
-    private func deleteVote(postId: Int) throws {
-        let deletedVote = votes.first { $0.post == postId }!
-        votes.removeAll { $0.id == deletedVote.id }
+    private func deletePostVote(postId: Int) throws {
+        guard let deletedVote = postVotes.first(where: { $0.post == postId } ) else { return }
+        postVotes.removeAll { $0.id == deletedVote.id }
         
         Task {
             do {
                 try await PostVoteAPI.deleteVote(voter: UserService.singleton.getId(), post: postId)
             } catch {
-                votes.append(deletedVote)
+                postVotes.append(deletedVote)
                 throw(error)
             }
         }
     }
+    
+    //MARK: - Comment Vote Updaters
+    
+    // Intermediate layer
+    func handleCommentVoteUpdate(commentId: Int, _ isAdding: Bool) throws {
+        if isAdding {
+            try castCommentVote(commentId: commentId)
+        } else {
+            try deleteCommentVote(commentId: commentId)
+        }
+    }
+    
+    private func castCommentVote(commentId: Int) throws {
+        let addedVote = CommentVote(id: Int.random(in: 0..<Int.max), voter: UserService.singleton.getId(), comment: commentId, timestamp: Date().timeIntervalSince1970, rating: nil)
+        commentVotes.append(addedVote)
+        
+        Task {
+            do {
+                let _ = try await CommentVoteAPI.postVote(voter: UserService.singleton.getId(), comment: commentId)
+            } catch {
+                commentVotes.removeAll { $0.id == addedVote.id }
+                throw(error)
+            }
+        }
+    }
+    
+    private func deleteCommentVote(commentId: Int) throws {
+        guard let deletedVote = commentVotes.first(where: { $0.comment == commentId } ) else { return }
+        commentVotes.removeAll { $0.id == deletedVote.id }
+        
+        Task {
+            do {
+                try await CommentVoteAPI.deleteVote(vote_id: deletedVote.id)
+            } catch {
+                commentVotes.append(deletedVote)
+                throw(error)
+            }
+        }
+    }
+    
+}
+
+
 //
 //    private func handleVoteAdd(postId: Int, emoji: String) throws {
 //        let addedVote = PostVote(id: Int.random(in: 0..<Int.max),
@@ -137,5 +187,3 @@ class VoteService: NSObject {
 //            }
 //        }
 //    }
-    
-}
